Develop

See merge request hieulv/training-server!4

+++ src/controllers/user.js
@@ -4,6 +4,7 @@ import { environment } from "../config/index.js";
 import { UserModel } from "../models/UserModel.js";
 import { generateRandomString } from "../utils/generateRandomString.js";
 import { sendMailCreateUser, validateEmail } from "../utils/handleEmail.js";
+import { DepartmentModel } from "../models/DepartmentModel.js";
 
 export const createUser = async (req, res) => {
   const { email, password, fullName, role, gender } = req.body;
@@ -254,6 +255,15 @@ export const deleteUser = async (req, res) => {
   const { arrayId } = req.body;
   try {
     for (let i = 0; i < arrayId.length; i++) {
+      const department = await DepartmentModel.findOne({
+        managerId: arrayId[i],
+      });
+      if (department) {
+        return res.status(200).json({
+          success: false,
+          message: "Can't be deleted, there is one or more relationship here",
+        });
+      }
       await UserModel.findByIdAndDelete({ _id: arrayId[i] });
     }
     res.status(200).json({
@@ -287,16 +297,16 @@ export const updateUser = async (req, res) => {
   try {
     const user = await UserModel.findOne({ _id: userId });
 
+    // my update
     if (userId === req.params.id) {
       await UserModel.findOneAndUpdate(
-        { _id: req.params.id },
+        { _id: userId },
         {
           fullName,
           dateOfBirth,
           gender,
           address,
           role,
-          department,
           joinCompanyAt,
           phone,
         }
@@ -310,20 +320,72 @@ export const updateUser = async (req, res) => {
       });
     }
 
+    // root update
     if (user.role === 0) {
-      await UserModel.findOneAndUpdate(
-        { _id: req.params.id },
-        {
-          fullName,
-          dateOfBirth,
-          gender,
-          address,
-          role,
-          department,
-          joinCompanyAt,
-          phone,
-        }
-      );
+      const oldUser = await UserModel.findOne({ _id: req.params.id });
+
+      // check update role for manager
+      // ex: role: 2 -> role: 1 true when department of manager: []
+      if (
+        oldUser.role === 1 &&
+        oldUser.department.length > 0 &&
+        oldUser.role !== role
+      ) {
+        return res.status(200).json({
+          success: false,
+          message: "Please update the department before updating this manger",
+        });
+      }
+
+      // if manager don't update department of manager
+      if (oldUser.role === 1) {
+        await UserModel.findOneAndUpdate(
+          { _id: req.params.id },
+          {
+            fullName,
+            dateOfBirth,
+            gender,
+            address,
+            role,
+            joinCompanyAt,
+            phone,
+          }
+        );
+      }
+
+      // update staff -> manager => department: []
+      if (oldUser.role === 2 && oldUser.role != role) {
+        await UserModel.findOneAndUpdate(
+          { _id: req.params.id },
+          {
+            fullName,
+            dateOfBirth,
+            gender,
+            address,
+            role,
+            department: [],
+            joinCompanyAt,
+            phone,
+          }
+        );
+      }
+
+      // if staff can update department of staff
+      if (oldUser.role === 2 && oldUser.role == role) {
+        await UserModel.findOneAndUpdate(
+          { _id: req.params.id },
+          {
+            fullName,
+            dateOfBirth,
+            gender,
+            address,
+            role,
+            department,
+            joinCompanyAt,
+            phone,
+          }
+        );
+      }
 
       const newUser = await UserModel.findOne({ _id: req.params.id });
       return res.status(200).json({
@@ -337,6 +399,7 @@ export const updateUser = async (req, res) => {
       message: "No permission to edit",
     });
   } catch (error) {
+    console.log("[ERROR UPDATE USER]]", error);
     res.status(200).json({ success: false, message: "Internal server error" });
   }
 };
@@ -348,16 +411,23 @@ export const viewUser = async (req, res) => {
   try {
     const userLogin = await UserModel.findOne({ _id: userId });
     const user = await UserModel.findOne({ _id: req.params.id });
+
     if (userLogin.role === 0)
       return res.status(200).json({
         success: true,
         message: "View user successfully",
         data: user,
       });
-    if (
-      userLogin.role === 1 &&
-      userLogin.department[0] === user.department[0]
-    ) {
+
+    const { department } = user;
+    let checked = false;
+    for (let i = 0; i < department?.length; i++) {
+      if (userLogin.department?.includes(department[i])) {
+        checked = true;
+      }
+    }
+
+    if (userLogin.role === 1 && checked) {
       return res.status(200).json({
         success: true,
         message: "View user successfully",
